fix linux compile error

+++ onnx/defs/schema.cc
@@ -716,7 +716,7 @@ OpSchema& OpSchema::AddQueriedFunctionBody(FunctionBodyQueryFunction queryFuncti
 }
 
 const FunctionProto* OpSchema::GetQueriedFunction(InferenceContext& ctx) const {
-  for (int i = 0; i < queried_function_bodies_.size(); i++) {
+  for (size_t i = 0; i < queried_function_bodies_.size(); i++) {
     if (queried_function_bodies_[i].first(ctx))
       return &queried_function_bodies_[i].second;
   }
@@ -745,11 +745,11 @@ OpSchema& OpSchema::FillUsing(const std::function<void(OpSchema&)>& populator) {
 // only add input to the function body if it is used by a body's node.
 bool FilterFunctionBodyInputByName(const std::string& input_name, const FunctionProto& function_body)
 {
-  for (size_t i = 0; i< function_body.node_size(); i++)
+  for (int i = 0; i< function_body.node_size(); i++)
   {
-    auto node = function_body.node((int)i);
-    for (size_t j = 0; j < node.input_size(); j++)
-      if (node.input((int)j) == input_name)
+    auto node = function_body.node(i);
+    for (int j = 0; j < node.input_size(); j++)
+      if (node.input(j) == input_name)
         return true;
   }
   return false;
